Update issue/PR templates with link to CONTRIBUTING.md
GitHub used to display a link to CONTRIBUTING.md at the top of the new issue/PR creation page but now they've changed to only showing a reminder to read it when the file has changed since user's last contribution. The link will only be clickable if the user switches to the Preview view, but I don't have a better solution.